requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
werkzeug==3.0.1
python-dotenv==1.0.0
openai==1.3.0

